auto_data_insights_generator/requirements.txt
# Django & REST
Django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.3
python-dotenv>=1.0
dj-database-url>=2.1
psycopg2-binary>=2.9

# Data Processing
pandas>=2.1
numpy>=1.26
openpyxl>=3.1

# Machine Learning & Analysis
scikit-learn>=1.3

# AI / NLP
google-generativeai>=0.3

# Report Generation
reportlab>=4.0

# Utilities
Pillow>=10.0
